+++ src/main/java/io/junrock/GAZA/domain/message/controller/MessageAlertController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
 
-import static io.junrock.GAZA.domain.message.dto.ButtonType.CAUTION;
+import static io.junrock.GAZA.global.type.ButtonType.CAUTION;
 
 
 @RestController
